refactor: extract action index validation
